Update PWM control demo for ESC PWM protocol
The commit updates the PWM control demo to command ESCs. It should
provide an implementation of the standard PWM signal described in

  https://quadmeup.com/pwm-oneshot125-oneshot42-and-multishot-comparison/

Our ESCs are documented as supporting this PWM protocol. This is not
the fastest ESC protocol our processor can support; however, it was
the easiest to implement and verify with a logic analyzer.

Note that pressing the SPACE bar sets the duty cycle to 0, not the
throttle percentage. SPACE still acts as a kill switch. But, we set
the throttle percentage to 0 during start-up.