Update project to use teensy4-bsp 0.1 release
I held off for a while on a formal BSP release. The BSP was using
some custom libraries that made it incompatible with the rest of
the embedded Rust ecosystem. Recent learnings have shown ways to
overcome those incompatibilities, so I was able to publish a
real release.

The PR updates all tools in this project to use the new, public
release of the BSP.

Tested by building the `pwm-control` demo, and flashing it on my
board.

https://crates.io/crates/teensy4-bsp
https://docs.rs/teensy4-bsp/0.1.0/teensy4_bsp/