Add pypwm_control Python package
The pypwm_control package provides a Python interface to the
pwm-control demo example for the embedded system. Use this to create
a host-side control loop.

Update the imu-parse.py example to use pypwm_control's ImuStream
abstraction. See the update tool for a demonstration of IMU reading
and filtering.

Tested the PwmControl interface from the new esc-throttle command-
line utility. Did not measure PWM outputs, but I observed that the
LED blinks at different frequencies, depending on the maximum
throttle.